crawler_consumer/django/consumer/requirements.txt
Django==3.1.5
PyYAML==5.3.1
ipython==7.19.0
jedi==0.17.2
mysqlclient==2.0.3
drf-yasg
django-filter
